refactoring

+++ domain/src/main/java/com/example/domain/models/Basket.kt
@@ -3,12 +3,12 @@ package com.example.domain.models
 import com.squareup.moshi.Json
 import com.squareup.moshi.JsonClass
 
+
 @JsonClass(generateAdapter = true)
 data class Basket(
-    val id: Long,
+    val id: Int,
     @Json(name = "images")
     val imageUrl: String,
     val price: Int,
     val title: String
 )
-
